refactor and simplify
