+++ agents/langchain/competitive-intelligence-agent/agent.py
