--- agents/langchain/competitive-intelligence-agent/ci_agent/__init__.py
@@ -1,5 +1,0 @@
-"""Competitive intelligence agent built on LangChain Deep Agents + Nebius Token Factory."""
-
-from ci_agent.agent import build_agent
-
-__all__ = ["build_agent"]
